refactor: split phase2 conformance helpers
